Fix: Add check on changed file
Once a file is detected as legal to be sent, it is prepared for transfer and will
be no more reused, until it is first deleted then recreated.

Since some issues arized while a new file replaced the old one, which is a wrong
way to use the FileMonitor, since there is no garanty of the previous transfer
(which might therefore end in error dur to mismatch file from out and in), a new
option is proposed: -ignoreAlreadyUsed (or <ignoreAlreadyUsed>true</ignoreAlreadyUsed>)

This option is false by default: the semantic is therefore changed by default but should
not make any difference on correct usage (which is: not replacing a file not yet sent,
using a post-action to delete or move the original file for instance).

However for users that do have this issue, the default setting will allow to have the
following new semantic:
- once the file is validated for transfer, the file is now taken into account however later on
if its modification time and hash is changing.
- once the modification is over, the file is reproposed to the R66 monitor
- if this was a 1 partner configuration model (file being sent to 1 partner only), the
previous transfer is requested to be canceled, and a new one is created; if there are
more than 1 partner, as the transfer id is unknown, the pending requests will probably
end up in error due to hash changing, but the new transfer will be ok

Reverting to previous semantic is simple: set -ignoreAlreadyUsed to true.

Additional changes:
- Fix missing tests on GatewayFTP (wrong .gitgnore)
- Fix dependencies (upgrade)