Fix: If no connection, retry is enhanced
While transfer might be in error (no connexion), Waarp R66
allow to retry after some delay (delayCommand in millisecond) the send.

If the transfer is in error or unknown, it is ignored, except if no connection is detected.
If the transfer is on going or done totally with no error, it will not be restarted.
If the source file is removed, the corresponding task is stopped in error with file not found error.

- Contains also a fix on ignoreAlreadyUsed option.
- Contains also a fix that detect early if source file is removed, therefore no send is possible (for send not self not through only)