Fix #17 JRE never exits when HostId not allowed
The JRE never exits when a DirectTransfer receives a "HostId not allowed" error #17

When such a message was raized, the current session was not checking if authentication
was over, therefore ignoring the wait and lock on this part.
Now it checks the missing Startup and Connection futures and if not set, set them in error.

Fix also fast fail on bad authentication with correct error at the end.