Fix issue #19 on missing DbRule for DbTaskRunner
When a DbRule is missing (deleted), one DbTaskRunner sibling it cannot be shown
in Https Admin interface. As this is obviously an issue to have such DbTaskRunner
with such missing DbRule, we fix however the printing such that this DbTaskRunner
is shown with its DbRule in RED color.

When a user asks to delete a Rule, now it is checked that no transfer exist with
such reference to this Rule before to act on deletion.
Note that this does not prevent full reload of Rules (delete all and reimport all),
meaning that extreme cases can occur where a rule needed is however deleted (so
the reason for the printing to take this however into account).

Note however that all transfers that need this Rule to exist continue to return an
error (such as on cancel, stop, restart operation).

Issue #19 is therefore fixed.