Fix issue #26 on DB Connection closed in Admin
In admin interface, espcially in Responsive mode, the DbConnection sometimes
seems closed and not reopened as it should be.
Since the new model of connections is to not keep the connections opens but to
reopen it as much as we need it (seems not to have big impact, and in particular
for Web Admin interface, it should not have any impact), this fix removes the
database keeping rule replacing by the underlying native pooling connections.

This fix is made for all Admin Interface: R66, GWFTP, ProxyR66.

This is only a fix for #26, not an evolution.