Fix issue on unlimited connection retry attempts
During fix on version 3.1/3.3, several issues concerning forwarding transfers
and FileMonitor introduce a endless repetition of attemps in case of lack
of connection.
This fixes this issue:

- After 3 attemps of failed connections, the transfer is in error and the
execution of the error task is launched
- For FileMonitor (SpooledDirectoryTransfer), a failed Submit transfer is handled
by the R66 server as usual (3 attempts). For Direct transfer, the retry after 3
attemps is done after a delay, as long as the file exists.