Fix issue on wrong argument on requester host
During change to DAO, one particular DbTaskRunner constructor
was wrongly modified, not taking into account the remote host
type (ssl or not) to get the local name of the host (again, ssl
or not).

This fix this issue, restablishing the original behavior.
Note that an existing method is marked as deprecated but kept
in order to allow upward compatiblity.

A test is added to prevent regression again.