Add option follow for transfer
This option `-follow` allows to add a specific field in the transfer information such that
in an externa tool, it is possible to link multiple transfers as hop of the same file.

For instance, sending a file from host A to host B, which will autoatically resend this file
to host C, and then to host D, all transfers will have a special information within the
transfer information such as `{'follow': number}`. This number will be kept the same
in each and every transfers related to the first one.

When specifiying the option `-follow`, for the first transfer, Waarp will allocate
a new specific Id of follow. Then all other transfers related to the first one will
have this one.