Add Rest V2 options for FollowId
The REST V2 interface integrates the search options based on `followId`
(`GET /v2/transfers/?followId=number`).

`number` being possibly a long number, it is adviced to manipulate
it as a String.

Add a special finder in DbTaskRunner for this
(`getSelectSameFollowId(followId, orderByStart, limit)`).

Fix also Map manipulation within Transfer Information to not loose other informations
than JSON extra information (various usages).

Note, for old versions, it is possible to simulate this option by doing the
following:
- Only for the first transfer, add to the `-info` transfer option a map as
`{"follow": uniqueNumber}` with a unique number such as nanotime or timestamp.
- To request on old versions, one should request all transfers (filtering on date
for instance) and post filtering them on the presence in the ``transferInfo` field
of such a Json map.