Fix FileInfo and TransferInfo
In some old revision, an inversion was done on FileInfo and TransferInfo.

FileInfo is the information given by the end user through `-info` and is linked to the file transfer.
This information contains the `FollowId` field (as Json Map) to allow to forward this info in retransfer.

TransferInfo is the internal information, stored by R66 servers and clients, into database about the transfer.
It could be the original size, some special information (such as Digest, UUID) but also the Follow ID.
Its internal representation is a JSON Map saved as a String (so escaped).

This fixes this partial inversion and also improves the HTML rendering by explicitly presenting the `Follow Id`.
This fixes also some cases where the Follow Id was ignored during retransfer (`TransferTask`).