Increasing concurrency limit to 10.000
The RUNLIMIT maximum value is turned to 10.000 while the default is still 1.000.

This allows huge config (more than 4 cores) to increase the concurrency capacity.
Note that the specified number (or default value) tends to allow up to 2 times
active or pending threads in the JVM.

Usually, a JVM thread limits is about 1.000 per core in practice.
So a 2 cores with Waarp will be able to use up to 2.000 threads so a limit of 1.000 for RUNLIMIT.
A 4 cores with Waarp will be able to use up to 4.000 threads so a limit of 2.000 for RUNLIMIT.
The new maximum value (10.000) will allow up to 20.000 threads, probably for a 20 cores server.