Add Compression tasks and by Packet
For Waarp R66:
- COMPRESS task is added using ZSTD fast and efficient compression.
- Compression can be used with compatible partners if compression is True
  in the configuration file (usefull for low bandwidth network)
- Add some options to REST V2 API using countOrder=true to get the number of
  items instead of full list.

Fix some issues on self transfers, such as Pre tasks not runned.

Fix interface of MonitorExporter to allow credentials for both API and Elasticsearch.

Optimizes buffer allocation with reuse and weak reference to allow garbage collecting.

The ZSTD implementation is based on zstd-jni and Aircompressor with a modified
version to be compatible with JRE 6 and future JRE without UNSAFE support.