Improve Compression specification
When compression is enabled in configuration file, to enable the compression per
block, the file transfer must have the following directive in the transfer information:

  `#COMPRESS#`

If the remote partner allows compression too, the compression will be used for this
transfer.
If the remote server does not allow compression (or is in a version priori to 3.6),
it will ignore tis directive and the compression will be not activated.

Fix also bytes to String and reverse conversion using UTF-8 as standard.

Fix documentation