Improve performance during authentication
Once a client is connected and authenticated, as long as the connection stay valid, the authentication process
could be skipped since the remote partner is already authenticated.
Note however that this optimization might not be compatible with R66 proxy based partner, since the connection might be
shared among several real partners over the R66 proxy.
To disable this, specify the option `-Dopenr66.authent.noreuse=1`

Also:
- Upgrade dependencies
- Fix bad init of some Benchmarks
- Fix bad check within tests
- Fix bad behavior on some functions where role has to be correct but then wrong exception occurs (create NoCorrectAuthentication)