Test on CPU Management are not stable enough
Doing test on CPU usage is valid only if the CI/CD servers are not loaded
by anything else. Nowadays, more of the CI/CD tools are not that calm.

SO the tests are changed to print an error but to not failed, since this is
not a real failure, just a status impacted by other not related programs
running at the same time than the tests.

Fix issue #104